requirements.txt
python-socketio[asyncio_client]>=5.11.0
aiohttp>=3.9.0
